+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [3.7.0] - 2026-06-26
+
 ### Added
 
 - Added `missing_module_policy` to `load_user_configs()` with `"error"`, `"warn"`, and `"ignore"` behaviors for optional modules.